Update folder architecture and correct small bug
Small bugs was sending two register request at first
breaking the connectionId
