--- awa-chat-flutter/awachat/lib/pages/slide.dart
@@ -1,93 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:url_launcher/url_launcher.dart';
-
-class SlideContainer extends StatelessWidget {
-  const SlideContainer({Key? key, required this.child}) : super(key: key);
-
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(24.0),
-      child: Center(
-        child: SingleChildScrollView(child: child),
-      ),
-    );
-  }
-}
-
-class Slide extends StatelessWidget {
-  const Slide({Key? key, required this.text, required this.assetPath})
-      : super(key: key);
-
-  final String text;
-  final String assetPath;
-  @override
-  Widget build(BuildContext context) {
-    return SlideContainer(
-        child: Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Image.asset(assetPath),
-        const Divider(height: 48),
-        Text(text, textAlign: TextAlign.center),
-      ],
-    ));
-  }
-}
-
-class CustomSlide extends StatelessWidget {
-  const CustomSlide({Key? key, required this.child, required this.assetPath})
-      : super(key: key);
-
-  final Widget child;
-  final String assetPath;
-  @override
-  Widget build(BuildContext context) {
-    return SlideContainer(
-        child: Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Image.asset(assetPath),
-        const Divider(height: 48),
-        child,
-      ],
-    ));
-  }
-}
-
-class SlideWithButton extends StatelessWidget {
-  const SlideWithButton(
-      {Key? key,
-      required this.text,
-      required this.assetPath,
-      required this.buttonText,
-      required this.onPressed})
-      : super(key: key);
-
-  final String text;
-  final String assetPath;
-  final String buttonText;
-  final VoidCallback onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return SlideContainer(
-        child: Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Image.asset(assetPath),
-        const Divider(height: 48),
-        Text(text, textAlign: TextAlign.center),
-        const Divider(height: 48),
-        ElevatedButton(
-            onPressed: onPressed,
-            child: Text(buttonText),
-            style: ElevatedButton.styleFrom(
-              primary: const Color(0xff6f61e8),
-            ))
-      ],
-    ));
-  }
-}
