+++ awa-chat-flutter/awachat/lib/widgets/agreements.dart
@@ -1,4 +1,4 @@
-import 'package:awachat/memory.dart';
+import 'package:awachat/objects/memory.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 
@@ -34,6 +34,10 @@ class Agreements extends StatelessWidget {
     );
   }
 }
+
+// ===== ===== =====
+// ===== ===== =====
+// HELPERS
 
 class AgreementsPage extends StatefulWidget {
   const AgreementsPage({
